% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ad.R
\name{ad_matrix_load_tape}
\alias{ad_matrix_load_tape}
\title{Load a saved matrix tape cache}
\usage{
ad_matrix_load_tape(path)
}
\arguments{
\item{path}{File created by `ADMatrixModel$save_tape()`.}
}
\value{
A recorded [ADMatrixModel].
}
\description{
Load a saved matrix tape cache
}
