% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ad_var}
\alias{ad_var}
\title{Create an AD variable manually}
\usage{
ad_var(name, value)
}
\arguments{
\item{name}{Variable name.}

\item{value}{Numeric scalar or vector value.}
}
\value{
A \code{Variable} R6 object.
}
\description{
Normally you do not need this function: \code{\link{backdiff}} converts
named parameter values to AD variables automatically. Use \code{ad_var} only
for advanced/manual graph construction.
}
\examples{
ad_var("x", 2)
}
