% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-mode.R
\name{backwardDifferentiate}
\alias{backwardDifferentiate}
\title{Reverse-mode backward pass (R implementation)}
\usage{
backwardDifferentiate(variable)
}
\arguments{
\item{variable}{Root node of the expression graph.}
}
\description{
Reverse-mode backward pass (R implementation)
}
\keyword{internal}
