% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Constant}
\alias{Constant}
\title{R6 class representing a constant node}
\format{
An \link[R6:R6Class]{R6::R6Class} generator object.
}
\description{
Constants participate in the graph but are not differentiated with respect to.
}
\examples{
c <- newConstant("c", 3)
c$value
}
