% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{newConstant}
\alias{newConstant}
\title{Create an AD constant}
\usage{
newConstant(name, value, grad = NULL)
}
\arguments{
\item{name}{Constant name.}

\item{value}{Numeric value.}

\item{grad}{Initial gradient (usually zero).}
}
\value{
A \code{Constant} R6 object.
}
\description{
Create an AD constant
}
\examples{
c <- newConstant("c", 3)
c$value
}
