% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{newVariable}
\alias{newVariable}
\title{Create an AD variable}
\usage{
newVariable(
  name,
  value,
  op = NULL,
  parents = list(),
  par = FALSE,
  expr = NULL,
  meta = NULL
)
}
\arguments{
\item{name}{Variable name used in symbolic expressions.}

\item{value}{Numeric value at which to evaluate.}

\item{op}{Internal operation label.}

\item{parents}{Internal parent nodes.}

\item{par}{Logical; mark as a differentiation parameter.}

\item{expr}{Internal parsed expression.}

\item{meta}{Internal metadata for some operations.}
}
\value{
A \code{Variable} R6 object.
}
\description{
Create an AD variable
}
\examples{
x <- newVariable("x", 2, par = TRUE)
x$value
}
