% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui-widget.R
\name{renderLibertadWorkbench}
\alias{renderLibertadWorkbench}
\title{Render a LibeRtAD workbench}
\usage{
renderLibertadWorkbench(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Widget expression.}

\item{env}{Evaluation environment.}

\item{quoted}{Whether expression is quoted.}
}
\description{
Render a LibeRtAD workbench
}
