% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad-state.R
\name{reset_tape}
\alias{reset_tape}
\title{Clears recorded operations and disables tape recording until re-enabled.}
\usage{
reset_tape()
}
\value{
Invisibly \code{NULL}.
}
\description{
Clears recorded operations and disables tape recording until re-enabled.
}
\examples{
f <- function(x) x^2
autodiff(f, x = 2, record_tape = TRUE)
reset_tape()
}
